+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@centeredgesoft/runtime-mock-routes",
-  "version": "2.0.2",
+  "version": "2.1.0",
   "description": "",
   "main": "build/index.js",
   "scripts": {
@@ -9,7 +9,7 @@
     "test": "jest"
   },
   "bin": {
-    "runtime-mock-routes": "./build/index.js"
+    "runtime-mock-routes": "./build/run.js"
   },
   "keywords": [],
   "author": "",
@@ -42,14 +42,14 @@
     "@types/faker": "^5.1.0",
     "@types/jest": "^26.0.14",
     "@types/lodash": "^4.14.161",
-    "@types/node": "^14.10.2",
+    "@types/node": "^14.11.2",
     "@types/supertest": "^2.0.10",
     "dotenv": "^8.2.0",
     "jest": "^26.4.2",
     "nodemon": "^2.0.4",
     "supertest": "^4.0.2",
-    "ts-jest": "^26.3.0",
+    "ts-jest": "^26.4.0",
     "ts-node": "^9.0.0",
-    "typescript": "^4.0.2"
+    "typescript": "^4.0.3"
   }
 }
