+++ tsconfig.json
@@ -4,6 +4,7 @@
         "module": "commonjs",
         "rootDir": "./",
         "outDir": "./build",
+        "declaration": true,
         "esModuleInterop": true,
         "strict": false,
     },
